JAVA-932: Remove EnumIntCodec and EnumStringCodec from public API.
This commit temporarily transfers these classes to the test directory. In the future, they could be part of a separate module for extra codecs, see JAVA-846.
