+++ driver-core/src/main/java/com/datastax/driver/core/TypeCodec.java
@@ -2121,109 +2121,4 @@ public abstract class TypeCodec<T> {
         }
     }
 
-    /**
-     * A codec that serializes {@link Enum} instances as CQL {@link DataType#varchar() varchar}s
-     * representing their programmatic names as returned by {@link Enum#name()}.
-     * <p>
-     * This codec is automatically used by the driver to create codecs for {@link Enum}
-     * instances on the fly.
-     *
-     * @param <E> The Enum class this codec serializes from and deserializes to.
-     */
-    public static class EnumStringCodec<E extends Enum<E>> extends TypeCodec<E> {
-
-        private final Class<E> enumClass;
-
-        private final TypeCodec<String> innerCodec;
-
-        public EnumStringCodec(Class<E> enumClass) {
-            this(VarcharCodec.instance, enumClass);
-        }
-
-        public EnumStringCodec(TypeCodec<String> innerCodec, Class<E> enumClass) {
-            super(innerCodec.getCqlType(), enumClass);
-            this.innerCodec = innerCodec;
-            this.enumClass = enumClass;
-        }
-
-        @Override
-        public ByteBuffer serialize(E value, ProtocolVersion protocolVersion) throws InvalidTypeException {
-            return value == null ? null : innerCodec.serialize(value.name(), protocolVersion);
-        }
-
-        @Override
-        public E deserialize(ByteBuffer bytes, ProtocolVersion protocolVersion) throws InvalidTypeException {
-            String value = innerCodec.deserialize(bytes, protocolVersion);
-            return value == null ? null : Enum.valueOf(enumClass, value);
-        }
-
-        @Override
-        public String format(E value) throws InvalidTypeException {
-            if (value == null)
-                return NULL;
-            return innerCodec.format(value.name());
-        }
-
-        @Override
-        public E parse(String value) throws InvalidTypeException {
-            return value == null || value.isEmpty() || value.equals(NULL) ? null : Enum.valueOf(enumClass, innerCodec.parse(value));
-        }
-
-    }
-
-    /**
-     * A codec that serializes {@link Enum} instances as CQL {@link DataType#cint() int}s
-     * representing their ordinal values as returned by {@link Enum#ordinal()}.
-     * <p>
-     * Note that, by default, the driver uses {@link EnumStringCodec} to serialize {@link Enum} instances;
-     * to force the use of this codec instead for a specific {@link Enum} instance,
-     * the appropriate {@link EnumIntCodec} instance must be explicitly
-     * {@link CodecRegistry#register(TypeCodec) registered}.
-     * <p>
-     * <strong>Note that this codec relies on the enum constants declaration order;
-     * it is therefore vital that this order remains immutable.</strong>
-     *
-     * @param <E> The Enum class this codec serializes from and deserializes to.
-     */
-    public static class EnumIntCodec<E extends Enum<E>> extends TypeCodec<E> {
-
-        private final E[] enumConstants;
-
-        private final TypeCodec<Integer> innerCodec;
-
-        public EnumIntCodec(Class<E> enumClass) {
-            this(IntCodec.instance, enumClass);
-        }
-
-        public EnumIntCodec(TypeCodec<Integer> innerCodec, Class<E> enumClass) {
-            super(innerCodec.getCqlType(), enumClass);
-            this.enumConstants = enumClass.getEnumConstants();
-            this.innerCodec = innerCodec;
-        }
-
-        @Override
-        public ByteBuffer serialize(E value, ProtocolVersion protocolVersion) throws InvalidTypeException {
-            return value == null ? null : innerCodec.serialize(value.ordinal(), protocolVersion);
-        }
-
-        @Override
-        public E deserialize(ByteBuffer bytes, ProtocolVersion protocolVersion) throws InvalidTypeException {
-            Integer value = innerCodec.deserialize(bytes, protocolVersion);
-            return value == null ? null : enumConstants[value];
-        }
-
-        @Override
-        public String format(E value) throws InvalidTypeException {
-            if (value == null)
-                return NULL;
-            return innerCodec.format(value.ordinal());
-        }
-
-        @Override
-        public E parse(String value) throws InvalidTypeException {
-            return value == null || value.isEmpty() || value.equals(NULL) ? null : enumConstants[innerCodec.parse(value)];
-        }
-
-    }
-
 }
